Support for different text domains
Both gettext and Babel make heavy use of what is called the text domain
or message domain [1]. This essentially is a unique name that identifies
the application, and it's also used to name the `pot`, `po` and `mo`
files. The code and the documentation currently assume that the domain
is always ``messages``, which is fine for a standalone application. But
when integrating with other applications or migrating old code bases you
might need to use a different domain.

This patch adds a new optional configuration option to define the
domain.

It also adds documentation for the BABEL_TRANSLATION_DIRECTORIES option
added on 11c7b29fe.

[1]
https://www.gnu.org/software/gawk/manual/html_node/Explaining-gettext.html