Add basic support for locales unkown by Babel
Right now Flask breaks if you request content in a locale that is
unknown by Babel even if the translation files actually exist.

This commit allows translations to be loaded even if the locale
is unknown by Babel. Date and number formats will fallback to
the app's default locale. Because of that, the default locale
is still required to be parseable by Babel.