Use flask.g as context
This is according to a recommendation on extension development in Flask
documentation. This also fixes an issue where force_locale context leaks
between two threads which are running in app context rather than request
context. `current_app` is a proxy to the application object, while
`flask.g` lives only for the duration of the context, either app context
or request context. `types.SimpleNamespace` is used with the extension
name as a prefix to avoid collisions.