Allow loading translations for a specific locale
So far, flask_babel supports translation to the locale selected for the
request. If one needs a translation into a different locale, one would
need to replicate parts of flask_babel's code. A use case for this is,
for example, sending an email to a site administrator in the
admininistrator's language. E.g.

    from flask_babel import Domain
    from babel import Locale

    domain = Domain(locale=Locale.parse(admin_lang))
    email_body = render_template('admin_email.html', _=domain.gettext)